# openapi.yaml
openapi: 3.0.3
info:
  title: FitLife API
  description: |
    REST API untuk aplikasi FitLife — platform nutrisi dan kesehatan berbasis Next.js.

    ## Autentikasi
    API ini mendukung dua metode autentikasi:
    1. **JWT Bearer Token** (direkomendasikan untuk Flutter/mobile) — kirim token di header `Authorization: Bearer <token>`
    2. **Cookie httpOnly** (untuk web browser) — cookie `token` di-set otomatis setelah login

    Token JWT berlaku selama **7 hari** dan berisi payload: `userId`, `role`, `email`.

    ## Base URL
    - **Lokal:** `http://localhost:3000`
    - **Produksi:** `https://<domain-anda>.vercel.app`
  version: "1.0.0"
  contact:
    name: FitLife Dev Team
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://your-domain.vercel.app
    description: Production server (Vercel)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token JWT diperoleh dari response login/register. Gunakan di header:
        `Authorization: Bearer <token>`
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Budi Santoso"
        username:
          type: string
          nullable: true
          example: "budi_s"
        email:
          type: string
          example: "budi@example.com"
        role:
          type: string
          enum: [user, admin]
          example: "user"
        phone:
          type: string
          nullable: true
          example: "08123456789"
        birthdate:
          type: string
          format: date-time
          nullable: true
          example: "1995-06-15T00:00:00.000Z"
        weight:
          type: integer
          nullable: true
          description: Berat badan dalam kg
          example: 70
        height:
          type: integer
          nullable: true
          description: Tinggi badan dalam cm
          example: 170
        photo:
          type: string
          nullable: true
          description: URL foto profil dari Cloudinary
          example: "https://res.cloudinary.com/..."
        google_avatar:
          type: string
          nullable: true
          description: URL avatar dari Google OAuth
        created_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
          nullable: true
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          example: "Budi Santoso"
        username:
          type: string
          example: "budi_santoso"
        phone:
          type: string
          example: "08123456789"
        birthdate:
          type: string
          format: date
          example: "1995-06-15"
        weight:
          type: integer
          example: 70
        height:
          type: integer
          example: 170
...
    UploadMenuResponse:
      type: object
      properties:
        url:
          type: string
          description: URL gambar yang sudah diupload ke Cloudinary
          example: "https://res.cloudinary.com/fitlife/image/upload/..."
    UploadProfileResponse:
      type: object
      properties:
        message:
          type: string
          example: "Foto berhasil diupload"
        user:
          $ref: '#/components/schemas/UserProfile'

    # ── Common ────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Unauthorized"
...
paths:
...
  /api/profile:
    get:
      tags: [Profile]
      summary: Ambil profil user yang login
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Data profil user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User tidak ditemukan

    patch:
      tags: [Profile]
      summary: Update profil user
      description: Update data profil. Semua field opsional — hanya field yang dikirim yang akan diupdate.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profil berhasil diperbarui
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Profil diperbarui"
                  user:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
        '409':
          description: Username sudah digunakan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Username sudah digunakan"
  /api/profile/upload:
    post:
      tags: [Profile]
      summary: Upload foto profil user
      description: |
        Upload foto profil ke Cloudinary. Format: `multipart/form-data`.
        Foto lama akan otomatis di-overwrite.

        **Batasan:**
        - Format: JPEG, JPG, PNG, WebP
        - Ukuran maksimal: **2MB**
        - Gambar akan di-crop ke 400×400 dengan fokus wajah
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photo]
              properties:
                photo:
                  type: string
                  format: binary
                  description: File gambar (JPEG/PNG/WebP, max 2MB)
      responses:
        '200':
          description: Foto berhasil diupload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadProfileResponse'
        '400':
          description: File tidak ada, tipe tidak didukung, atau ukuran terlalu besar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
...
  /api/upload/menu:
    post:
      tags: [Upload]
      summary: Upload gambar menu ke Cloudinary
      description: |
        Upload gambar untuk menu makanan. Gambar akan di-resize ke 800×600.

        **Batasan:**
        - Format: JPEG, JPG, PNG, WebP
        - Ukuran maksimal: **3MB**

        **Membutuhkan autentikasi.**
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: File gambar (JPEG/PNG/WebP, max 3MB)
                folder:
                  type: string
                  description: Folder Cloudinary (opsional)
                  default: "fitlife/menus"
                  example: "fitlife/menus"
      responses:
        '200':
          description: Gambar berhasil diupload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadMenuResponse'
              example:
                url: "https://res.cloudinary.com/fitlife/image/upload/..."
        '400':
          description: File tidak ada, tipe tidak didukung, atau terlalu besar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error